Halcyon CSI
Updated the script. Went back to the original isocenter calculation placement. Added in an equal distance calculator to equally place additional isocenters between the brain and final isocenter. I'm pretty happy overall with the script. May look at changing the x-coordinate placement to automatically place it in the middle of the brain structure as well.